Extract Msf::DBManager::Import::CI
MSP-11124
